+++ src/util/EntityToFormMapper.js
@@ -13,6 +13,16 @@ import evaluator from './helpers/evaluator'
 // $FlowFixMe
 import api from '@molgenis/molgenis-api-client'
 
+const DEFAULTS = {
+  mapperMode: 'UPDATE',
+  booleanLabels: {
+    trueLabel: 'True',
+    falseLabel: 'False',
+    nillLabel: 'N/A'
+  },
+  showNillableBooleanOption: true
+}
+
 // Create an object type UserException
 function MappingException (message: string) {
   this.message = message
@@ -33,7 +43,6 @@ MappingException.prototype.toString = function () {
  *
  * @param refEntity The refEntity of the attribute.
  * @param search An optional search query used to filter the items of the response
- * @param multiple An optional boolean specifying that the search value is an array of values
  * @return {Promise} Promise object representing an Array of FieldOption
  */
 const fetchFieldOptions = (refEntity: RefEntityType, search: ?string | ?Array<string>): Promise<Array<FieldOption>> => {
@@ -98,17 +107,11 @@ const fetchFieldOptions = (refEntity: RefEntityType, search: ?string | ?Array<st
  * @param options MapperOptions optional object containing options to configure mapper
  * @returns {Function|null} Function which returns a Promise representing an Array of FieldOptions
  */
-const getFieldOptions = (attribute, options?: MapperOptions): ?(() => Promise<Array<FieldOption>>) => {
+const getFieldOptions = (attribute, options: MapperSettings): ?(() => Promise<Array<FieldOption>>) => {
   const fetchOptionsFunction = (search: ?string | Array<string>): Promise<Array<FieldOption>> => {
     return fetchFieldOptions(attribute.refEntity, search).then(response => {
       return response
     })
-  }
-
-  const booleanLabels = {
-    trueLabel: options && options.booleanLabels ? options.booleanLabels.trueLabel : 'True',
-    falseLabel: options && options.booleanLabels ? options.booleanLabels.falseLabel : 'False',
-    nillLabel: options && options.booleanLabels ? options.booleanLabels.nillLabel : 'N/A'
   }
 
   switch (attribute.fieldType) {
@@ -142,14 +145,12 @@ const getFieldOptions = (attribute, options?: MapperOptions): ?(() => Promise<Ar
       return (): Promise<Array<FieldOption>> => Promise.resolve(enumOptions)
     case 'BOOL':
       const boolOptions = [
-        {id: 'true', value: true, label: booleanLabels.trueLabel},
-        {id: 'false', value: false, label: booleanLabels.falseLabel}
+        {id: 'true', value: true, label: options.booleanLabels.trueLabel},
+        {id: 'false', value: false, label: options.booleanLabels.falseLabel}
       ]
 
-      if (attribute.nillable &&
-        (!options || !options.hasOwnProperty('showNillableBooleanOption') || options.showNillableBooleanOption)
-      ) {
-        boolOptions.push({id: 'null', value: null, label: booleanLabels.nillLabel})
+      if (attribute.nillable && options.showNillableBooleanOption) {
+        boolOptions.push({id: 'null', value: null, label: options.booleanLabels.nillLabel})
       }
 
       return (): Promise<Array<FieldOption>> => Promise.resolve(boolOptions)
@@ -266,23 +267,42 @@ const entityTypeToSubFieldType = (entityFieldType: EntityFieldType): SubType | n
 }
 
 /**
+ * Determine if field should be disabled
+ * @param attribute
+ * @param mapperOptions
+ * @returns boolean
+ */
+const isDisabledField = (attribute, mapperOptions: MapperSettings): boolean => {
+  if (attribute.fieldType === 'ONE_TO_MANY') {
+    return true
+  }
+
+  if (mapperOptions.mapperMode === 'CREATE') {
+    return false
+  }
+
+  return attribute.readOnly
+}
+
+/**
  * Generate a schema field object suitable for the forms
  *
  * @param attribute Attribute metadata from an EntityType V2 response
  * @param mapperOptions MapperOptions optional object containing options to configure mapper
  * @returns {{type: String, id, label, description, required: boolean, disabled, visible, options: ({uri, id, label, multiple}|{uri, id, label})}}
  */
-const generateFormSchemaField = (attribute, mapperOptions?: MapperOptions): FormField => {
+const generateFormSchemaField = (attribute, mapperOptions: MapperSettings): FormField => {
   // options is a function that always returns an array of option objects
   const options = getFieldOptions(attribute, mapperOptions)
+  const isDisabled = isDisabledField(attribute, mapperOptions)
   let fieldProperties = {
     id: attribute.name,
     label: attribute.label,
     description: attribute.description,
     type: getHtmlFieldType(attribute.fieldType),
     required: isRequired(attribute),
-    disabled: attribute.readOnly || attribute.fieldType === 'ONE_TO_MANY',
-    readOnly: attribute.readOnly || attribute.fieldType === 'ONE_TO_MANY',
+    disabled: isDisabled,
+    readOnly: isDisabled,
     visible: isVisible(attribute),
     validate: isValid(attribute)
   }
@@ -355,29 +375,73 @@ const generateFormData = (fields: any, data: any, attributes: any) => {
 }
 
 /**
+ * Returns true if entity attribute should be included in form
+ *
+ * Some attributes do not map to a actionable from item ( for example a server side generated auto id field)
+ *
+ * @param attribute
+ * @returns {boolean}
+ */
+const isFormFieldAttribute = (attribute: any):boolean => {
+  return !(attribute.auto && !attribute.visible)
+}
+
+/**
  * Generates an array for form fields
  *
  * @param attributes A list of MOLGENIS attribute metadata
- * @param options MapperOptions optional object containing options to configure mapper
+ * @param options MapperOptions object containing options to configure mapper
  * @returns a an array of Field objects
  */
-const generateFormFields = (attributes: any, options?: MapperOptions): Array<FormField> => attributes.reduce((accumulator, attribute) => {
-  accumulator.push(generateFormSchemaField(attribute, options))
-  return accumulator
-}, [])
+const generateFormFields = (attributes: any, options: MapperSettings): Array<FormField> => attributes
+  .filter(isFormFieldAttribute)
+  .map((attr) => {
+    return generateFormSchemaField(attr, options)
+  })
+
+/**
+ * Construct mapper settings taking into account the user settings, if no settings are passed the defaults are used
+ * @param settings
+ * @returns {{mapperMode: *, booleanLabels: {trueLabel: string, falseLabel: string, nillLabel: string}, showNillableBooleanOption: boolean}}
+ */
+const buildMapperSettings = (settings?: MapperOptions): MapperSettings => {
+  if (!settings) {
+    return DEFAULTS
+  }
+
+  const mapperMode = settings.mapperMode ? settings.mapperMode : DEFAULTS.mapperMode
+  let booleanLabels = DEFAULTS.booleanLabels
+  if (settings.booleanLabels) {
+    booleanLabels = {
+      trueLabel: settings.booleanLabels.trueLabel ? settings.booleanLabels.trueLabel : 'True',
+      falseLabel: settings.booleanLabels.falseLabel ? settings.booleanLabels.falseLabel : 'False',
+      nillLabel: settings.booleanLabels.nillLabel ? settings.booleanLabels.nillLabel : 'N/A'
+    }
+  }
+  let showNillableBooleanOption = DEFAULTS.showNillableBooleanOption
+  if (typeof (settings.showNillableBooleanOption) === 'boolean') {
+    showNillableBooleanOption = settings.showNillableBooleanOption
+  }
+
+  return {
+    mapperMode: mapperMode,
+    booleanLabels: booleanLabels,
+    showNillableBooleanOption: showNillableBooleanOption
+  }
+}
 
 /**
  * Generates both fields and data objects for rendering a form
  *
  * @param metadata MOLGENIS metadata, containing attributes
  * @param data
- * @param options MapperOptions optional object containing options to configure mapper
+ * @param userSettings MapperOptions optional object containing options to configure mapper
  * @returns {{formFields: Array<FormField>, formData: *}}
  */
-const generateForm = (metadata: any, data: ?any, options?: MapperOptions) => {
+const generateForm = (metadata: any, data: ?any, userSettings?: MapperOptions) => {
+  const mapperOptions = buildMapperSettings(userSettings)
   const attributes = metadata.attributes
-
-  const formFields = generateFormFields(attributes, options)
+  const formFields = generateFormFields(attributes, mapperOptions)
   const formData = data ? generateFormData(formFields, data, attributes) : {}
 
   return {